Update robotmetrics.py
In line 122, result_file_name variable has been used by using this it was creating a metrics file in root directory instead of creating the file as we mentioned as inputpath argument. so by changing the result_file_name with result_file, it is creating the metrics file at the same location as inputpath argument.

Please review my changes and let me know if my suggestion is wrong as this is my first analysis and change in  github. Thanks.